+++ src/resonaate/data/task.py
@@ -6,7 +6,7 @@ from sqlalchemy import Boolean, Column, Float, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
 # Local Imports
-from . import Base, _DataMixin
+from .table_base import Base, _DataMixin
 
 
 class Task(Base, _DataMixin):